README.md
Name: Generosity 
Authors: Tauqir Abdullah & Ihtesham Chowdhury

Description:
This app is designed to be fun while also teaching the theme of generosity. The objective is to get as many Heaven points as possible. You start out by working hard and clicking on the coin to earn money. Eventually you begin to invest it on Earth to earn more. However, as you earn more, you should always think about others and so at that point you must donate money that you earn to gain Heaven points. If you don't donate something for a while, your rate of gain for Heaven points starts to decay. 

We first designed the app and its functions using a google flowchart:
https://docs.google.com/drawings/d/1M384KD-70t5cFrf_aOpt07l7hFjU7VhsbJIIvAQ6ogo/edit

From there, we moved forward with programming different parts of the app. All images were created in Photoshop.

Thanks for playing!